pyproject.toml
[tool.poetry]
name = "nonebot-plugin-hotsearch"
version = "0.2.1"
description = "nonebot2 plugin for hotsearch"
authors = ["Astolfocat <2364934266@qq.com>"]
readme = "README.md"
packages = [{include = "nonebot_plugin_hotsearch"}]
license = "MIT"


[tool.poetry.dependencies]
python = ">=3.9,<4.0"
nonebot2 = ">=2.0.0"
nonebot-adapter-onebot = ">=2.1.3"
httpx = ">=0.23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
